Update gamelog to v0.7.0
