+++ gamelog/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gamelog"
-version = "0.6.0"
+version = "0.7.0"
 edition = "2024"
 
 [dependencies]
