+++ gamelog/src/lib.rs
@@ -9,7 +9,7 @@ mod play;
 mod terrain;
 
 #[allow(unused)]
-pub const MIN_VER: semver::Version = semver::Version::new(0, 6, 0);
+pub const MIN_VER: semver::Version = semver::Version::new(0, 7, 0);
 
 // I'm lazy.
 pub use action::*;
